# docs/swagger.yaml
definitions:
  app_cart_biz_model.CartItem:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  app_checkout_biz_model.Address:
    properties:
      city:
        type: string
      country:
        type: string
      state:
        type: string
      street_address:
        type: string
      zip_code:
        type: string
    type: object
  app_checkout_biz_model.CreditCardInfo:
    properties:
      credit_card_cvv:
        type: integer
      credit_card_expiration_month:
        type: integer
      credit_card_expiration_year:
        type: integer
      credit_card_number:
        type: string
    type: object
  app_order_biz_model.Address:
    properties:
      city:
        type: string
      country:
        type: string
      state:
        type: string
      street_address:
        type: string
      zip_code:
        type: integer
    type: object
  app_order_biz_model.CartItem:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  app_payment_biz_model.CreditCardInfo:
    properties:
      credit_card_cvv:
        type: integer
      credit_card_expiration_month:
        type: integer
      credit_card_expiration_year:
        type: integer
      credit_card_number:
        type: string
    type: object
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  model.AddItemReqCopy:
    properties:
      item:
        $ref: '#/definitions/app_cart_biz_model.CartItem'
      user_id:
        type: integer
    type: object
  model.ChargeReqCopy:
    properties:
      amount:
        type: number
      credit_card:
        $ref: '#/definitions/app_payment_biz_model.CreditCardInfo'
      order_id:
        type: string
      user_id:
        type: integer
    type: object
  model.CheckoutReq:
    properties:
      address:
        $ref: '#/definitions/app_checkout_biz_model.Address'
      credit_card:
        $ref: '#/definitions/app_checkout_biz_model.CreditCardInfo'
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      user_id:
        type: integer
    type: object
  model.EmptyCartReqCopy:
    properties:
      user_id:
        type: integer
    type: object
  model.GetCartReqCopy:
    properties:
      user_id:
        type: integer
    type: object
  model.GetProductReq:
    properties:
      id:
        type: integer
    type: object
  model.ListOrderReq:
    properties:
      user_id:
        type: integer
    type: object
  model.ListProductsReq:
    properties:
      categoryName:
        type: string
      page:
        type: integer
      pageSize:
        type: integer
    type: object
  model.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.MarkOrderPaidReq:
    properties:
      order_id:
        type: string
      user_id:
        type: integer
    type: object
  model.OrderItem:
    properties:
      cost:
        type: number
      item:
        $ref: '#/definitions/app_order_biz_model.CartItem'
    type: object
  model.PlaceOrderReq:
    properties:
      address:
        $ref: '#/definitions/app_order_biz_model.Address'
      email:
        type: string
      order_items:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      user_currency:
        type: string
      user_id:
        type: integer
    type: object
  model.RegisterReq:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
  /cart:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除的购物车信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.EmptyCartReqCopy'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 删除购物车api
      tags:
      - 购物车服务
    get:
      consumes:
      - application/json
      parameters:
      - description: 查询的购物车信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.GetCartReqCopy'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 查询购物车api
      tags:
      - 购物车服务
    post:
      consumes:
      - application/json
      parameters:
      - description: 新增的购物车信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.AddItemReqCopy'
      produces:
      - application/json
      responses:
        "200":
          description: 新增成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 新增购物车api
      tags:
      - 购物车服务
  /checkout:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建的结算信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.CheckoutReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 创建结算记录api
      tags:
      - 结算服务
  /order:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 标记的订单信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.MarkOrderPaidReq'
      produces:
      - application/json
      responses:
        "200":
          description: 标记成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 标记订单为已完成api
      tags:
      - 订单服务
    get:
      consumes:
      - application/json
      parameters:
      - description: 查询的订单信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ListOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 批量查询订单api
      tags:
      - 订单服务
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建的订单信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.PlaceOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 创建订单api
      tags:
      - 订单服务
  /payment:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建的支付信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ChargeReqCopy'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 创建支付记录api
      tags:
      - 支付服务
  /product:
    get:
      consumes:
      - application/json
      parameters:
      - description: 查询的商品 id
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.GetProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 根据 id 查询商品api
      tags:
      - 商品服务
  /product/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 查询的商品和分页信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ListProductsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 分页查询商品api
      tags:
      - 商品服务
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录的用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 登录用户api
      tags:
      - 用户服务
  /user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 新增的用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/common.Response'
      summary: 注册用户api
      tags:
      - 用户服务
swagger: "2.0"